backend/requirements.txt
uvicorn
fastapi
#for template  #new
jinja2
#for static files
aiofiles
#for database   #new
sqlalchemy
psycopg2-binary
#for loading environment variables  #new
python-dotenv
#for email validation            #new
pydantic[email]
#hashing              #new
passlib[bcrypt]
#for testing       #new
pytest
requests
#for jwt tokens    #new
python-jose
#for  cleaning up codebase on each commit
pre-commit

#for accessing form data       #new
python-multipart

#for google connexion
authlib
itsdangerous
httpx
